refactor: Favorite관련 Listener를 하나로 통합
